+++ game-shared-types/src/index.ts
@@ -1,3 +1,0 @@
-// src/index.ts
-export * from './game.types';
-export * from './client.types';